Split PostgreSql repo in seperate classes
