+++ molgenis-data-postgresql/src/main/java/org/molgenis/data/postgresql/PostgreSqlConfiguration.java
@@ -2,10 +2,7 @@ package org.molgenis.data.postgresql;
 
 import javax.sql.DataSource;
 
-import org.molgenis.data.DataService;
 import org.molgenis.data.ManageableRepositoryCollection;
-import org.molgenis.data.elasticsearch.IndexedManageableRepositoryCollectionDecorator;
-import org.molgenis.data.elasticsearch.SearchService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -16,22 +13,16 @@ import org.springframework.jdbc.core.JdbcTemplate;
 public class PostgreSqlConfiguration
 {
 	@Autowired
-	private DataService dataService;
-
-	@Autowired
 	private PostgreSqlEntityFactory postgreSqlEntityFactory;
 
 	@Autowired
 	private DataSource dataSource;
 
-	@Autowired
-	private SearchService searchService;
-
 	@Bean
 	@Scope("prototype")
 	public PostgreSqlRepository postgreSqlRepository()
 	{
-		return new PostgreSqlRepository(dataService, postgreSqlEntityFactory, jdbcTemplate());
+		return new PostgreSqlRepository(postgreSqlEntityFactory, jdbcTemplate());
 	}
 
 	@Bean
@@ -58,6 +49,6 @@ public class PostgreSqlConfiguration
 				throw new UnsupportedOperationException();
 			}
 		};
-		return new IndexedManageableRepositoryCollectionDecorator(searchService, postgreSqlRepositoryCollection);
+		return postgreSqlRepositoryCollection;
 	}
 }
